\name{cap-class}
\Rdversion{1.1}
\docType{class}
\alias{cap-class}
\title{Class \code{"cap"} in Package \pkg{HandTill2001}}
\description{
  A virtual class for \code{"bincap"} and  \code{"multcap"}.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{response}:}{Object of class \code{"factor"}, typically two class or multiple class observations.}
  }
}
\section{Methods}{
No methods defined with class "cap" in the signature.
}
\author{
  Andreas Dominik Cullmann
}
\seealso{
  \code{"\link[=bincap-class]{class?bincap}"}, \code{"\link[=multcap-class]{class?multcap}"}
}
\keyword{classes}
